Add snapshot methods for snapshotting an animation via a non-element view
These snapshot methods can be used for snapshotting an animation using a view other than the element being animated. This makes it possible to snapshot animations for which the element type is not a view, such as in the added tests where a proxy object is being animated.

They can also be used to snapshot a container view, when the animation modifies the element itself rather than subelements. For example, if an animation included keyframes for its element's `transform`, the snapshots should be of a view above the element in the view hierarchy to visualize the change.