Update handling of execution and per-frame execution blocks in child animations
This makes a few updates to the handling of execution blocks and per-frame execution blocks in child animations. Previously, per-frame execution blocks in child animations were only executed during the duration where the child was added, but the progress values in the context corresponded to that of the parent animation (#14). This breaks down the ability for child animations to independently reason about phases of the animation without context on the parent animation.

A similar problem existed with execution blocks, where the relative timestamp at which the block is executed was relative to the uncurved progress. This broke down when a child animation was being collapsed into the parent.

In order to address these issues, the following changes were made:

* Per-frame execution blocks are executed based on the curved progress of the parent animation, while the context shows the progress of the child animation.

* The timestamp for execution blocks is determined based off of the curved progress of the animation.

* The state of the animation instance is updated before the frame is rendered, to enable per-frame execution blocks to check the progress of the overall parent animation.

There is an argument to be made for basing the execution blocks off of the uncurved progress, but this can be achieved by wrapping the curved animation in a linear animation and adding the execution blocks to that parent.