Delay calling completion handlers until final frame is drawn to screen
Previously, completion handlers for an animation were called in the same display loop pass as the rendering of the final frame. If the completion handlers did any significant amount of work, this could result in delayed drawing of the final frame.

This moves the call to the completion handlers to happen on the next display loop pass after the final frame is rendering, to ensure that the final frame is drawn to screen without delay.