Add animation curve snapshot tests
This adds snapshot tests for a handful of animation curves, with snapshots that show the curve drawn on a grid.

The code to draw the grid is broken out into a separate utility shared with the sample app.

This is a combination of nine commits:

Checkpoint: add files

Fix demo app build

Cleanup

Cleanup

Cleanup

Add iPhone 7 10.3.1 snapshots

Add iPhone X 11.4 and 12.4 snapshots

Add iPhone X 13.3 snapshots

Cleanup