Update CI badge on README to show GitHub Actions
The CI badge was previously showing the status of the travis-ci.org build. This build has since been disabled with the move to travis-ci.com, so the badge wasn't showing the current build status.

This updates the badge to show the status of the GitHub Actions CI build for master.

Alternatively, we could show two badges: one for GitHub Actions and one for Travis (using the travis-ci.com status). This would accurately represent the fact that our CI is split across the two systems, but I think showing one badge is cleaner and we've moved most of the builds to GitHub Actions, with Travis only handling legacy iOS and Xcode versions, so I think the GitHub Actions status is more important to show.