Remove the collection keyframes feature
This feature was never included in the public API and still had a lot of work left in order to get it working properly (in particular around interacting with other animation content). No progress has been made on this feature in the last 9 months and AFAIK there is no work planned for this in the near future.

This removes it from the framework for now to simplify things.