Fix pod lint command for StagehandTesting
Previous, this was using `--external-podspecs` to tell the linter to use the local podspec for Stagehand. While this did successfully point it at the local podspec, it still used the tag to clone the source.

This switches over to using `--include-podspecs`, which tells it to use both the local podspec and the local source.